host/pyproject.toml
[project]
name = "cameracommander"
version = "0.1.0"
description = "Headless host application for the CameraCommander2 motorised pan-tilt head: timelapse and video-pan orchestration with REST + WebSocket API."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.30.1",
    "pydantic>=2.7.4",
    "typer>=0.12.3",
    "pyyaml>=6.0.1",
    "pyserial>=3.5",
    "gphoto2>=2.5.0",
    "pillow>=10.3.0",
    "piexif>=1.1.3",
    "loguru>=0.7.2",
]

[project.scripts]
cameracommander = "cameracommander.cli.main:app"

[tool.uv]
dev-dependencies = [
    "pytest>=8.2.2",
    "pytest-asyncio>=0.23.7",
    "httpx>=0.27.0",
    "ruff>=0.4.10",
    "mypy>=1.10.1",
    "types-pyyaml>=6.0.12",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
addopts = "-ra --strict-markers"
markers = [
    "integration: end-to-end tests against the mock firmware + mock camera",
    "contract: contract / schema-conformance tests",
    "slow: tests that exceed 1 s",
]
